Rework plot() and associated notebook
Rework code in plot() to allow for unit conversion when effective radius is one of the things being plotted. Also clean up plot() calls in MSR notebook to make the data clearer. Note this rework may not be complete. Based on the plots in the MSR notebook I suspect changes will need to be made to the new private method reConvert() to ensure the correct data is being used for photometric vs spectroscopic redshifts.