Progress on bootstrapping
Lines created and plotted correctly. 68% confidence intervals plotted correctly. Still a hangup when trying to draw the two curves over the endpoints of the interval lines. Tried getting the y-values of the lines from the ErrorBarContainer object but it only has the median (input y) values. Need to find another method for getting these points.